requirements/core.txt
# MeshForge Core Dependencies (required for all profiles)

# Terminal formatting (TUI)
rich>=13.0.0

# Configuration and utilities
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
distro>=1.8.0
